requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
authlib==1.6.0
bcrypt==4.3.0
billiard==4.2.1
cachetools==5.5.2
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.1.1
cryptography==45.0.3
dnspython==2.7.0
docstring-parser==0.16
email-validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
fastapi-limiter==0.1.6
google-adk==1.2.1
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.96.0
google-cloud-appengine-logging==1.6.1
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-speech==2.32.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.1
google-crc32c==1.7.1
google-genai==1.19.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.20.3
grpc-google-iam-v1==0.14.2
grpcio==1.72.1
grpcio-status==1.72.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib-metadata==8.7.0
jinja2==3.1.6
kombu==5.5.4
loguru==0.7.3
markdown-it-py==3.0.0
markupsafe==3.0.2
mcp==1.9.3
mdurl==0.1.2
nodeenv==1.9.1
numpy==2.3.0
opentelemetry-api==1.34.0
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.34.0
opentelemetry-semantic-conventions==0.55b0
packaging==25.0
prisma==0.15.0
prompt-toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-core==2.33.2
pydantic-settings==2.9.1
pygments==2.19.1
pyjwt==2.10.1
pymongo==4.13.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pyyaml==6.0.2
redis==6.2.0
requests==2.32.3
resend==2.10.0
rich==14.0.0
rich-toolkit==0.14.7
rsa==4.9.1
ruff==0.11.12
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.41
sse-starlette==2.3.6
starlette==0.46.2
tomlkit==0.13.2
typer==0.16.0
typing-extensions==4.13.2
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
zipp==3.23.0
